runtime: support capturing C backtrace from signal handler on darwin/amd64
The implementation is mostly copied from the commit that added
linux/amd64 for this feature (ea306ae).

I plan to build a Darwin equivalent of github.com/ianlancetaylor/cgosymbolizer,
but I wanted to get this proposal out for review first. Is this patch
something the Go team is interested in supporting?